Move parameter JSON generation from Model to schema generator
